[fix] testcode refactor

+++ src/main/java/io/pp/arcade/v1/domain/user/dto/UserDetailResponseDto.java
@@ -25,7 +25,7 @@ public class UserDetailResponseDto {
         this.level = level;
         this.currentExp = currentExp;
         this.maxExp = maxExp;
-        this.expRate = (double) (currentExp / maxExp * 100);
+        this.expRate = ((double)currentExp / (double)maxExp) * 100;
         this.rivalRecord = rivalRecord;
     }
